+++ packages/create-sitecore-jss/src/templates/angular-xmcloud/src/app/components/page-content/page-content.component.ts
@@ -7,6 +7,11 @@ import { Subscription } from 'rxjs';
 @Component({
   selector: 'app-page-content',
   templateUrl: './page-content.component.html',
+  host: {
+    'class': 'component content',
+    '[class]': 'styles',
+    '[id]': 'id',
+  }
 })
 export class PageContentComponent extends SxaComponent implements OnInit, OnDestroy {
   content?: RichTextField;