+++ packages/create-sitecore-jss/src/templates/node-xmcloud-proxy/src/index.ts
@@ -3,6 +3,7 @@ import express, { Response } from 'express';
 import compression from 'compression';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import { debug } from '@sitecore-jss/sitecore-jss';
+import { editingRouter } from '@sitecore-jss/sitecore-jss-proxy';
 import { config } from './config';
 
 const server = express();
@@ -22,10 +23,11 @@ const requiredProperties = [
   'parseRouteUrl',
   'clientFactory',
   'getClientFactoryConfig',
-  'siteName',
   'defaultLanguage',
   'layoutServiceFactory',
   'dictionaryServiceFactory',
+  'components',
+  'metadata',
 ];
 
 const missingProperties = requiredProperties.filter((property) => !config.serverBundle[property]);
@@ -38,15 +40,19 @@ if (missingProperties.length > 0) {
   );
 }
 
+const layoutService = layoutServiceFactory.create();
+
+const dictionaryService = dictionaryServiceFactory.create();
+
+const clientFactoryConfig = config.serverBundle.getClientFactoryConfig();
+
 /**
  * GraphQL endpoint resolution to meet the requirements of the http-proxy-middleware
  */
 const graphQLEndpoint = (() => {
-  const clientFactoryConfig = config.serverBundle.getClientFactoryConfig();
-
   try {
     const graphQLEndpoint = new URL(clientFactoryConfig.endpoint);
-    // Browser request path to the proxy. Includes only the pathname. 
+    // Browser request path to the proxy. Includes only the pathname.
     const pathname = graphQLEndpoint.pathname;
     // Target URL for the proxy. Can't include the query string.
     const target = `${graphQLEndpoint.protocol}//${graphQLEndpoint.hostname}${pathname}`;
@@ -62,10 +68,6 @@ const graphQLEndpoint = (() => {
     );
   }
 })();
-
-const layoutService = layoutServiceFactory.create();
-
-const dictionaryService = dictionaryServiceFactory.create();
 
 /**
  * Parse requested url in order to detect current route and language
@@ -122,6 +124,23 @@ server.use(
   createProxyMiddleware({
     target: graphQLEndpoint.target,
     changeOrigin: true,
+  })
+);
+
+/**
+ * Proxy editing requests through the editing router
+ */
+server.use(
+  '/api/editing',
+  editingRouter({
+    config: {
+      components: config.serverBundle.components,
+      metadata: config.serverBundle.metadata,
+    },
+    render: {
+      clientFactory: config.serverBundle.clientFactory,
+      renderView,
+    },
   })
 );