+++ packages/sitecore-jss-nextjs/src/editing/editing-render-middleware.test.ts
@@ -8,6 +8,9 @@ import { EditingDataService, EditingPreviewData } from './editing-data-service';
 import {
   EDITING_ALLOWED_ORIGINS,
   QUERY_PARAM_EDITING_SECRET,
+  RenderMetadataQueryParams,
+} from '@sitecore-jss/sitecore-jss/editing';
+import {
   QUERY_PARAM_VERCEL_PROTECTION_BYPASS,
   QUERY_PARAM_VERCEL_SET_BYPASS_COOKIE,
 } from './constants';
@@ -15,7 +18,6 @@ import { EE_PATH, EE_LANGUAGE, EE_LAYOUT, EE_DICTIONARY, EE_BODY } from '../test
 import {
   ChromesHandler,
   EditingRenderMiddleware,
-  MetadataQueryParams,
   isEditingMetadataPreviewData,
 } from './editing-render-middleware';
 import { spy, match } from 'sinon';
@@ -36,7 +38,7 @@ const allowedOrigin = 'https://allowed.com';
 
 const mockRequest = (
   body?: any,
-  query?: Query | MetadataQueryParams,
+  query?: Query | RenderMetadataQueryParams,
   method?: string,
   headers?: { [key: string]: string }
 ) => {
@@ -179,7 +181,8 @@ describe('EditingRenderMiddleware', () => {
       sc_variant: 'dev',
       sc_version: 'latest',
       secret: secret,
-    } as MetadataQueryParams;
+      sc_layoutKind: 'shared',
+    } as RenderMetadataQueryParams;
 
     it('should handle request', async () => {
       const req = mockRequest(EE_BODY, query, 'GET');
@@ -198,6 +201,7 @@ describe('EditingRenderMiddleware', () => {
         version: 'latest',
         editMode: 'metadata',
         pageState: 'edit',
+        layoutKind: 'shared',
       });
 
       expect(res.redirect).to.have.been.calledOnce;
@@ -217,7 +221,7 @@ describe('EditingRenderMiddleware', () => {
         sc_site: 'website',
         secret: secret,
         sc_variant: 'id-1,id-2,id-3',
-      } as MetadataQueryParams;
+      } as RenderMetadataQueryParams;
 
       const req = mockRequest(EE_BODY, query, 'GET');
       const res = mockResponse();
@@ -235,6 +239,7 @@ describe('EditingRenderMiddleware', () => {
         version: undefined,
         editMode: 'metadata',
         pageState: 'edit',
+        layoutKind: undefined,
       });
     });
 
@@ -246,7 +251,7 @@ describe('EditingRenderMiddleware', () => {
         sc_lang: 'en',
         sc_site: 'website',
         secret: secret,
-      } as MetadataQueryParams;
+      } as RenderMetadataQueryParams;
       const req = mockRequest(EE_BODY, queryWithoutOptionalParams, 'GET');
       const res = mockResponse();
 
@@ -263,6 +268,7 @@ describe('EditingRenderMiddleware', () => {
         version: undefined,
         editMode: 'metadata',
         pageState: 'edit',
+        layoutKind: undefined,
       });
 
       expect(res.redirect).to.have.been.calledOnce;
@@ -295,6 +301,7 @@ describe('EditingRenderMiddleware', () => {
         version: 'latest',
         editMode: 'metadata',
         pageState: 'edit',
+        layoutKind: 'shared',
       });
 
       expect(res.redirect).to.have.been.calledOnce;