+++ packages/sitecore-jss-react/src/components/EditingScripts.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { EditMode, LayoutServicePageState } from '@sitecore-jss/sitecore-jss/layout';
 import { useSitecoreContext } from '../enhancers/withSitecoreContext';
+import { getJssPagesClientData } from '@sitecore-jss/sitecore-jss/editing';
 
 /**
  * Renders client scripts and data for editing/preview mode in Pages.
@@ -15,20 +16,21 @@ export const EditingScripts = (): JSX.Element => {
   if (pageState === LayoutServicePageState.Normal) return <></>;
 
   if (editMode === EditMode.Metadata) {
+    const jssClientData = { ...clientData, ...getJssPagesClientData() };
+
     return (
       <>
         {clientScripts?.map((src, index) => (
           <script src={src} key={index} />
         ))}
-        {clientData &&
-          Object.keys(clientData).map((id) => (
-            <script
-              key={id}
-              id={id}
-              type="application/json"
-              dangerouslySetInnerHTML={{ __html: JSON.stringify(clientData[id]) }}
-            />
-          ))}
+        {Object.keys(jssClientData).map((id) => (
+          <script
+            key={id}
+            id={id}
+            type="application/json"
+            dangerouslySetInnerHTML={{ __html: JSON.stringify(jssClientData[id]) }}
+          />
+        ))}
       </>
     );
   }