+++ packages/sitecore-jss/src/editing/index.ts
@@ -7,6 +7,10 @@ export {
   resetEditorChromes,
   handleEditorAnchors,
   Metadata,
+  getJssPagesClientData,
+  EDITING_ALLOWED_ORIGINS,
+  QUERY_PARAM_EDITING_SECRET,
+  PAGES_EDITING_MARKER,
 } from './utils';
 export {
   DefaultEditFrameButton,
@@ -19,3 +23,5 @@ export {
   EditButtonTypes,
   mapButtonToCommand,
 } from './edit-frame';
+export { RenderMetadataQueryParams } from './models';
+export { LayoutKind } from './models';